chore: added tests for common package. refactor: moved interpreter package.
