+++ docker/build-release.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+
 apt update && apt install -y zip
 
 cd /data
