+++ internal/core/core_test.go
