+++ internal/core/dependency.go
