+++ internal/core/preprocessor.go
