+++ internal/core/preprocessor_dependency.go
