+++ internal/core/preprocessor_foreach.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"errors"
 
-	"github.com/xiroxasx/yatt/internal/interpreter/foreach"
+	"github.com/xiroxasx/yatt/internal/foreach"
 )
 
 func (c *Core) foreachStart(pd *PreprocessorDirective) (err error) {
