+++ internal/core/testdata/deps/fileA.txt
