+++ internal/core/testdata/deps/fileB.txt
