+++ internal/core/testdata/deps/fileC.txt
