+++ internal/core/testdata/deps/fileD.txt
