+++ internal/core/testdata/deps/fileG.txt
