+++ internal/core/testdata/foreach/in/foreach.yaml
