+++ internal/core/testdata/functions/test.txt
