+++ internal/core/testdata/ignore/in/ignore.yaml
