+++ internal/core/testdata/imports/in/myTestImportFail.txt
