+++ internal/core/testdata/imports/in/myTestImportOk.txt
