+++ internal/core/testdata/imports/in/nestedImportOk.txt
