+++ internal/core/testdata/imports/in/startFail.txt
