+++ internal/core/testdata/vars/in/start.yaml
