+++ internal/core/testdata/vars/in/yatt.var
