+++ internal/core/variable.go
