+++ internal/foreach/prepare.go
