+++ internal/functions/crypt.go
