+++ internal/functions/internal.go
