+++ internal/functions/math.go
