+++ internal/functions/string.go
