+++ internal/functions/time.go
